+++ launchdarkly/resource_launchdarkly_custom_role_test.go
@@ -23,6 +23,8 @@ const (
 		}
 	}
 `
+	// IMPORTANT TO NOTE that the $ character must be escaped in terraform by using a double $$
+	// otherwas ${} will be interpreted as a terraform variable and throw an error
 	testAccCustomRoleUpdate = `
 resource "launchdarkly_custom_role" "test" {
 	key = "%s"
@@ -30,7 +32,7 @@ resource "launchdarkly_custom_role" "test" {
 	policy {
 		actions = ["*"]	
 		effect = "allow"
-		resources = ["proj/*:env/staging"]
+		resources = ["proj/*:env/$${roleAttribute/devEnvironments}"]
 	}
 }
 `
@@ -42,19 +44,7 @@ resource "launchdarkly_custom_role" "test" {
 	policy_statements {
 		actions = ["*"]	
 		effect = "allow"
-		resources = ["proj/*:env/staging"]
-	}
-}
-`
-	testAccCustomRoleCreateWithNotStatements = `
-resource "launchdarkly_custom_role" "test" {
-	key = "%s"
-	name = "Custom role - %s"
-	description = "Don't allow all actions on non-staging environments"
-	policy_statements {
-		not_actions = ["*"]	
-		effect = "allow"
-		not_resources = ["proj/*:env/staging"]
+		resources = ["proj/$${roleAttribute/devProjects}:env/staging"]
 	}
 }
 `
@@ -67,6 +57,18 @@ resource "launchdarkly_custom_role" "test" {
 		actions = ["*"]	
 		effect = "deny"
 		resources = ["proj/*:env/production"]
+	}
+}
+`
+	testAccCustomRoleCreateWithNotStatements = `
+resource "launchdarkly_custom_role" "test" {
+	key = "%s"
+	name = "Custom role - %s"
+	description = "Don't allow all actions on non-staging environments"
+	policy_statements {
+		not_actions = ["*"]	
+		effect = "allow"
+		not_resources = ["proj/*:env/staging"]
 	}
 }
 `
@@ -84,7 +86,7 @@ resource "launchdarkly_custom_role" "test" {
 `
 )
 
-func TestAccCustomRole_Create(t *testing.T) {
+func TestAccCustomRole_CreateAndUpdate(t *testing.T) {
 	key := acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
 	name := acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
 	resourceName := "launchdarkly_custom_role.test"
@@ -110,11 +112,27 @@ func TestAccCustomRole_Create(t *testing.T) {
 					resource.TestCheckResourceAttr(resourceName, "policy.0.effect", "deny"),
 				),
 			},
+			{
+				Config: fmt.Sprintf(testAccCustomRoleUpdate, key, name),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckCustomRoleExists(resourceName),
+					resource.TestCheckResourceAttr(resourceName, KEY, key),
+					resource.TestCheckResourceAttr(resourceName, NAME, "Updated - "+name),
+					resource.TestCheckResourceAttr(resourceName, DESCRIPTION, ""), // should be empty after removal
+					resource.TestCheckResourceAttr(resourceName, BASE_PERMISSIONS, "reader"),
+					resource.TestCheckResourceAttr(resourceName, "policy.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "policy.0.actions.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "policy.0.actions.0", "*"),
+					resource.TestCheckResourceAttr(resourceName, "policy.0.resources.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "policy.0.resources.0", "proj/*:env/${roleAttribute/devEnvironments}"),
+					resource.TestCheckResourceAttr(resourceName, "policy.0.effect", "allow"),
+				),
+			},
 		},
 	})
 }
 
-func TestAccCustomRole_CreateWithStatements(t *testing.T) {
+func TestAccCustomRole_CreateAndUpdateWithStatements(t *testing.T) {
 	key := acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
 	name := acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
 	resourceName := "launchdarkly_custom_role.test"
@@ -136,8 +154,29 @@ func TestAccCustomRole_CreateWithStatements(t *testing.T) {
 					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.actions.#", "1"),
 					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.actions.0", "*"),
 					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.resources.#", "1"),
-					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.resources.0", "proj/*:env/staging"),
+					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.resources.0", "proj/${roleAttribute/devProjects}:env/staging"),
 					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.effect", "allow"),
+				),
+			},
+			{
+				ResourceName:      resourceName,
+				ImportState:       true,
+				ImportStateVerify: true,
+			},
+			{
+				Config: fmt.Sprintf(testAccCustomRoleUpdateWithStatements, key, name),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckCustomRoleExists(resourceName),
+					resource.TestCheckResourceAttr(resourceName, KEY, key),
+					resource.TestCheckResourceAttr(resourceName, NAME, "Updated role - "+name),
+					resource.TestCheckResourceAttr(resourceName, DESCRIPTION, "Deny all actions on production environments"),
+					resource.TestCheckResourceAttr(resourceName, "policy.#", "0"),
+					resource.TestCheckResourceAttr(resourceName, "policy_statements.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.actions.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.actions.0", "*"),
+					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.resources.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.resources.0", "proj/*:env/production"),
+					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.effect", "deny"),
 				),
 			},
 			{
@@ -149,7 +188,7 @@ func TestAccCustomRole_CreateWithStatements(t *testing.T) {
 	})
 }
 
-func TestAccCustomRole_CreateWithNotStatements(t *testing.T) {
+func TestAccCustomRole_CreateAndUpdateWithNotStatements(t *testing.T) {
 	key := acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
 	name := acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
 	resourceName := "launchdarkly_custom_role.test"
@@ -181,98 +220,6 @@ func TestAccCustomRole_CreateWithNotStatements(t *testing.T) {
 				ImportState:       true,
 				ImportStateVerify: true,
 			},
-		},
-	})
-}
-
-func TestAccCustomRole_Update(t *testing.T) {
-	key := acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
-	name := acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
-	resourceName := "launchdarkly_custom_role.test"
-	resource.ParallelTest(t, resource.TestCase{
-		PreCheck: func() {
-			testAccPreCheck(t)
-		},
-		Providers: testAccProviders,
-		Steps: []resource.TestStep{
-			{
-				Config: fmt.Sprintf(testAccCustomRoleCreate, key, name),
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckCustomRoleExists(resourceName),
-				),
-			},
-			{
-				Config: fmt.Sprintf(testAccCustomRoleUpdate, key, name),
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckCustomRoleExists(resourceName),
-					resource.TestCheckResourceAttr(resourceName, KEY, key),
-					resource.TestCheckResourceAttr(resourceName, NAME, "Updated - "+name),
-					resource.TestCheckResourceAttr(resourceName, DESCRIPTION, ""), // should be empty after removal
-					resource.TestCheckResourceAttr(resourceName, BASE_PERMISSIONS, "reader"),
-					resource.TestCheckResourceAttr(resourceName, "policy.#", "1"),
-					resource.TestCheckResourceAttr(resourceName, "policy.0.actions.#", "1"),
-					resource.TestCheckResourceAttr(resourceName, "policy.0.actions.0", "*"),
-					resource.TestCheckResourceAttr(resourceName, "policy.0.resources.#", "1"),
-					resource.TestCheckResourceAttr(resourceName, "policy.0.resources.0", "proj/*:env/staging"),
-					resource.TestCheckResourceAttr(resourceName, "policy.0.effect", "allow"),
-				),
-			},
-		},
-	})
-}
-
-func TestAccCustomRole_UpdateWithStatements(t *testing.T) {
-	key := acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
-	name := acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
-	resourceName := "launchdarkly_custom_role.test"
-	resource.ParallelTest(t, resource.TestCase{
-		PreCheck: func() {
-			testAccPreCheck(t)
-		},
-		Providers: testAccProviders,
-		Steps: []resource.TestStep{
-			{
-				Config: fmt.Sprintf(testAccCustomRoleCreate, key, name),
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckCustomRoleExists(resourceName),
-				),
-			},
-			{
-				Config: fmt.Sprintf(testAccCustomRoleUpdateWithStatements, key, name),
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckCustomRoleExists(resourceName),
-					resource.TestCheckResourceAttr(resourceName, KEY, key),
-					resource.TestCheckResourceAttr(resourceName, NAME, "Updated role - "+name),
-					resource.TestCheckResourceAttr(resourceName, DESCRIPTION, "Deny all actions on production environments"),
-					resource.TestCheckResourceAttr(resourceName, "policy.#", "0"),
-					resource.TestCheckResourceAttr(resourceName, "policy_statements.#", "1"),
-					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.actions.#", "1"),
-					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.actions.0", "*"),
-					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.resources.#", "1"),
-					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.resources.0", "proj/*:env/production"),
-					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.effect", "deny"),
-				),
-			},
-		},
-	})
-}
-
-func TestAccCustomRole_UpdateWithNotStatements(t *testing.T) {
-	key := acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
-	name := acctest.RandStringFromCharSet(10, acctest.CharSetAlphaNum)
-	resourceName := "launchdarkly_custom_role.test"
-	resource.ParallelTest(t, resource.TestCase{
-		PreCheck: func() {
-			testAccPreCheck(t)
-		},
-		Providers: testAccProviders,
-		Steps: []resource.TestStep{
-			{
-				Config: fmt.Sprintf(testAccCustomRoleCreateWithStatements, key, name),
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckCustomRoleExists(resourceName),
-				),
-			},
 			{
 				Config: fmt.Sprintf(testAccCustomRoleUpdateWithNotStatements, key, name),
 				Check: resource.ComposeTestCheckFunc(
@@ -288,6 +235,11 @@ func TestAccCustomRole_UpdateWithNotStatements(t *testing.T) {
 					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.not_resources.0", "proj/*:env/production"),
 					resource.TestCheckResourceAttr(resourceName, "policy_statements.0.effect", "deny"),
 				),
+			},
+			{
+				ResourceName:      resourceName,
+				ImportState:       true,
+				ImportStateVerify: true,
 			},
 		},
 	})